refactor compound select
